upgrade-to-ses3: populate metadata heap with pool name
SES2 did not contain a meatadata_heap structure as part of the zone
configuration, and hence had no associated pool. After upgrade to SES3,
user creation is impossible as some metadata is meant to be written to the
non-existent pool. We address this here by, for each zone, populating the
zone config with the proper metadata_heap pool, which RADOSGW will create
on restart.

Fixes: SUSE/ses-upgrade-helper#16

Signed-off-by: Karol Mroz <kmroz@suse.com>