Refrain from touching CRUSH tunables (bsc#980683)
While the default CRUSH tunables in SES2.1 are bobtail and need to be changed
at some point before (or after) upgrading to SES3, the upgrade script is not
the place to do this.

Any change in the tunables potentially causes lots of data movement.  Also,
even on a small cluster you only run "ceph osd crush tunables optimal" once,
whereas in this case we were contemplating running it on every node.

Fixes: #8
Signed-off-by: Nathan Cutler <ncutler@suse.com>