upgrade-ses: user_ceph_not_in_use re-work for SES4
Originally this global preflight ensured that user 'ceph' was not in
use. This was needed as SES2 created the user 'ceph' as the admin user,
and the upgrade script would later rename this user to 'cephadm' so that
Ceph could utilize user and group 'ceph' for running daemons.

In the case of upgrading from SES3+, user 'ceph' in use is valid
provided it is the new ceph user (ie. belonging to group 'ceph').

Signed-off-by: Karol Mroz <kmroz@suse.de>