upgrade-ses: check OSD journal and data partitions
Add a global pre-flight to alert admin of invalid OSD journal or data
partition types.  OSD directories that are detected to be
invalid/inactive should be treated as warnings.  Invalid journal or data
partition types should be fatal, however, allow the user to skip this check
if they deem it necessary.

Fixes: bsc#994867

Signed-off-by: Karol Mroz <kmroz@suse.de>